Move catalog tests into separate files so they can be parallelized.
