+++ package.json
@@ -7,7 +7,8 @@
     "version": "chmod +x ./scripts/verify-node-version.sh && ./scripts/verify-node-version.sh && npm run docs && git add ./docs/*",
     "check-lisc": "license-checker --production --onlyAllow 'Apache-2.0;BSD-3-Clause;MIT'",
     "lint": "eslint 'src/**/*.js' 'spec/**/*.js'",
-    "test": "mkdir -p test && cp -rf src/* test && mocha ./spec/*",
+    "test": "mkdir -p test && cp -rf src/* test && mocha --parallel ./spec/*",
+    "test:sync": "mkdir -p test && cp -rf src/* test && mocha ./spec/*",
     "precoverage": "rm -rf ./coverage && rm -rf ./.nyc_output",
     "coverage": "nyc --all --reporter=html npm run test",
     "postcoverage": "open coverage/index.html && rm -rf test",
